Move get_active_shares and get_share_state
- Move from functions in state to methods of MetadataStore.
- Adjust test code to match.
